+++ source/ProcessManager.Orchestrations.Tests/Integration/Processes/BRS_021/ForwardMeteredData/V1/MonitorOrchestrationUsingClientsScenario.cs
@@ -13,10 +13,12 @@
 // limitations under the License.
 
 using Energinet.DataHub.Core.DurableFunctionApp.TestCommon.DurableTask;
+using Energinet.DataHub.Core.FunctionApp.TestCommon.EventHub.ListenerMock;
 using Energinet.DataHub.ProcessManager.Abstractions.Api.Model.OrchestrationInstance;
 using Energinet.DataHub.ProcessManager.Client;
 using Energinet.DataHub.ProcessManager.Client.Extensions.DependencyInjection;
 using Energinet.DataHub.ProcessManager.Client.Extensions.Options;
+using Energinet.DataHub.ProcessManager.Components.Extensions.Options;
 using Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.Processes.BRS_021.ForwardMeteredData.V1.Model;
 using Energinet.DataHub.ProcessManager.Orchestrations.Tests.Fixtures;
 using Energinet.DataHub.ProcessManager.Shared.Tests.Fixtures.Extensions;
@@ -63,6 +65,7 @@ public class MonitorOrchestrationUsingClientsScenario : IAsyncLifetime
     {
         _fixture.ProcessManagerAppManager.AppHostManager.ClearHostLog();
         _fixture.OrchestrationsAppManager.AppHostManager.ClearHostLog();
+        _fixture.EventHubListener.Reset();
 
         return Task.CompletedTask;
     }
@@ -79,7 +82,7 @@ public class MonitorOrchestrationUsingClientsScenario : IAsyncLifetime
     public async Task ForwardMeteredData_WhenStartedUsingClient_CanMonitorLifecycle()
     {
         // Arrange
-        var input = CreateMeteredDataForMeasurementPointMessageInputV1();
+        var input = CreateMeteredDataForMeteringPointMessageInputV1();
 
         var startCommand = new StartForwardMeteredDataCommandV1(
             new ActorIdentityDto(input.AuthenticatedActorId),
@@ -101,11 +104,12 @@ public class MonitorOrchestrationUsingClientsScenario : IAsyncLifetime
         var completedOrchestration = await _fixture.DurableClient.WaitForOrchestrationCompletedAsync(
             orchestration.InstanceId);
         completedOrchestration.RuntimeStatus.Should().Be(OrchestrationRuntimeStatus.Completed);
+        await _fixture.EventHubListener.WhenAny().VerifyOnceAsync();
     }
 
-    private static MeteredDataForMeasurementPointMessageInputV1 CreateMeteredDataForMeasurementPointMessageInputV1()
+    private static MeteredDataForMeteringPointMessageInputV1 CreateMeteredDataForMeteringPointMessageInputV1()
     {
-        var input = new MeteredDataForMeasurementPointMessageInputV1(
+        var input = new MeteredDataForMeteringPointMessageInputV1(
             Guid.NewGuid(),
             "EGU9B8E2630F9CB4089BDE22B597DFA4EA5",
             "571313101700011887",
@@ -114,35 +118,35 @@ public class MonitorOrchestrationUsingClientsScenario : IAsyncLifetime
             "K3",
             "2024-12-03T08:00:00Z",
             "PT1H",
-            "2024-12-01T23:00:00Z",
+            "2024-12-01T23:00Z",
             "2024-12-02T23:00:00Z",
             "5790002606892",
             null,
             new List<EnergyObservation>()
             {
-                new("1", "112.000", "E01"),
-                new("2", "112.000", "E01"),
-                new("3", "112.000", "E01"),
-                new("4", "112.000", "E01"),
-                new("5", "112.000", "E01"),
-                new("6", "112.000", "E01"),
-                new("7", "112.000", "E01"),
-                new("8", "112.000", "E01"),
-                new("9", "112.000", "E01"),
-                new("10", "112.000", "E01"),
-                new("12", "112.000", "E01"),
-                new("12", "112.000", "E01"),
-                new("13", "112.000", "E01"),
-                new("14", "112.000", "E01"),
-                new("15", "112.000", "E01"),
-                new("16", "112.000", "E01"),
-                new("18", "112.000", "E01"),
-                new("19", "112.000", "E01"),
-                new("20", "112.000", "E01"),
-                new("21", "112.000", "E01"),
-                new("22", "112.000", "E01"),
-                new("23", "112.000", "E01"),
-                new("24", "112.000", "E01"),
+                new("1", "112.000", "A04"),
+                new("2", "112.000", "A04"),
+                new("3", "112.000", "A04"),
+                new("4", "112.000", "A04"),
+                new("5", "112.000", "A04"),
+                new("6", "112.000", "A04"),
+                new("7", "112.000", "A04"),
+                new("8", "112.000", "A04"),
+                new("9", "112.000", "A04"),
+                new("10", "112.000", "A04"),
+                new("12", "112.000", "A04"),
+                new("12", "112.000", "A04"),
+                new("13", "112.000", "A04"),
+                new("14", "112.000", "A04"),
+                new("15", "112.000", "A04"),
+                new("16", "112.000", "A04"),
+                new("18", "112.000", "A04"),
+                new("19", "112.000", "A04"),
+                new("20", "112.000", "A04"),
+                new("21", "112.000", "A04"),
+                new("22", "112.000", "A04"),
+                new("23", "112.000", "A04"),
+                new("24", "112.000", "A04"),
             });
         return input;
     }