+++ source/ProcessManager.Orchestrations/Processes/BRS_021/ForwardMeteredData/V1/Orchestration_Brs_021_ForwardMeteredData_V1.cs
@@ -41,7 +41,7 @@ internal class Orchestration_Brs_021_ForwardMeteredData_V1
     public async Task<string> Run(
         [OrchestrationTrigger] TaskOrchestrationContext context)
     {
-        var input = context.GetOrchestrationParameterValue<MeteredDataForMeasurementPointMessageInputV1>();
+        var input = context.GetOrchestrationParameterValue<MeteredDataForMeteringPointMessageInputV1>();
 
         if (input == null)
             return "Error: No input specified.";
@@ -74,8 +74,10 @@ internal class Orchestration_Brs_021_ForwardMeteredData_V1
 
         // Step: Storing
         await context.CallActivityAsync(
-            nameof(StoreMeteredDataForMeasurementPointActivity_Brs_021_ForwardMeteredData_V1),
-            new StoreMeteredDataForMeasurementPointActivity_Brs_021_ForwardMeteredData_V1.ActivityInput(instanceId),
+            nameof(StoreMeteredDataForMeteringPointActivity_Brs_021_ForwardMeteredData_V1),
+            new StoreMeteredDataForMeteringPointActivity_Brs_021_ForwardMeteredData_V1.ActivityInput(
+                instanceId,
+                input),
             _defaultRetryOptions);
         //await context.WaitForExternalEvent<string>("Measurements_Notification");
         await context.CallActivityAsync(