+++ source/ProcessManager.Orchestrations/Processes/BRS_028/V1/Activities/EnqueueActorMessagesActivity_Brs_028_V1.cs
@@ -57,6 +57,7 @@ internal class EnqueueActorMessagesActivity_Brs_028_V1(
 
         return _enqueueActorMessagesClient.Enqueue(
             Orchestration_Brs_028_V1.Name,
+            input.InstanceId.Value,
             enqueuedBy.ToDto(),
             "enqueue-" + input.InstanceId.Value,
             acceptedData);