+++ docs/ProcessManager.Client/ReleaseNotes/ReleaseNotes.md
@@ -1,10 +1,17 @@
 # ProcessManager.Client Release Notes
 
-## Version 0.18.0
+## Version 0.19.0
 
 - Updated code documenation (XML comments).
 - Refactored implementation of custom queries support.
 - Implemented interface `IOrchestrationInstanceTypedDto<out TInputParameterDto>` and added it to `OrchestrationInstanceTypedDto<TInputParameterDto>`.
+
+## Version 0.18.0
+
+- Add version suffix to service bus messages names:
+    - `EnqueueActorMessagesV1`
+    - `StartOrchestrationV1`
+- Refactor `EnqueueActorMessagesV1` and `StartOrchestrationV1` service bus message fields
 
 ## Version 0.17.0